Move OpInfo subclasses into opinfo.core
[ghstack-poisoned]
